+++ app/api/game/guess/route.ts
@@ -1,86 +1,13 @@
-import words from "@/data/words";
 import { NextRequest } from "next/server";
 import {
 	decodeCandidates,
 	encodeCadidates,
+	findCandidates,
 	getCandidates,
 	getSession,
+	makeGuess,
+	validateGuess,
 } from "../utils";
-import { LetterMatch } from "@/lib/wordle/types";
-
-const findCandidates = (
-	guess: string,
-	candidates: { word: string; position: number }[]
-) => {
-	let selectedCandidates: { word: string; position: number }[] = [];
-	let minPoints = 50; // set to maximum points - H = 10, P = 1, M = 0
-
-	// if minPoints is updated, then clear candidates, and add the new item that made minPoints change
-	candidates.forEach((candidate) => {
-		const result = makeGuess(guess, candidate.word);
-		let points = 0;
-		for (let i = 0; i <= result.length; i++) {
-			const letter = result[i];
-			if (letter === "H") points += 10;
-			else if (letter === "P") points += 1;
-		}
-		if (points === minPoints) {
-			selectedCandidates.push(candidate);
-		} else if (points < minPoints) {
-			selectedCandidates = [candidate];
-			minPoints = points;
-		}
-	});
-
-	return { candidates: selectedCandidates, points: minPoints };
-};
-
-const makeGuess = (guess: string, answer: string) => {
-	const result = Array<LetterMatch>(guess.length).fill(LetterMatch.Miss);
-	const answerCharsTaken = Array<boolean>(answer.length).fill(false);
-
-	// check for correct positions
-	for (let i = 0; i < guess.length; i++) {
-		if (guess[i] === answer[i]) {
-			result[i] = LetterMatch.Hit;
-			answerCharsTaken[i] = true;
-		}
-	}
-
-	// check for wrong positions but correct letters
-	for (let i = 0; i < guess.length; i++) {
-		if (result[i] === LetterMatch.Miss) {
-			// Only check those not already marked as 'H'
-			for (let j = 0; j < answer.length; j++) {
-				if (!answerCharsTaken[j] && guess[i] === answer[j]) {
-					result[i] = LetterMatch.Present; // Correct letter, wrong position
-					answerCharsTaken[j] = true;
-					break; // Stop once a match is found
-				}
-			}
-		}
-	}
-
-	return result.join("");
-};
-
-type ValidateGuessResult = {
-	guess?: string;
-	error?: string;
-};
-
-const validateGuess = (guess: unknown): ValidateGuessResult => {
-	if (typeof guess !== "string") return { error: "guess must be a string" };
-
-	if (guess.length !== 5)
-		return { guess, error: "guess must be 5 characters long" };
-
-	const normalizedGuess = guess.toLowerCase();
-	if (!words.includes(normalizedGuess))
-		return { guess, error: "guess must be a valid word" };
-
-	return { guess: normalizedGuess };
-};
 
 export async function POST(request: NextRequest) {
 	const session = await getSession();
